Use short TTL for DNS lookups in combination with a local cache.
Unifies how DNS queries are cached by implementing a proper caching
mechanism with recursion detection.

The recursion detection is needed when the cluster's DNS server is
running on the local host (e.g. a k3s in Docker) since it may forward
queries that cannot resolve to the local DNS resolver, which in turn
forwards the query to Telepresence DNS resolver.

The number of queries that are forwarded to the cluster is minimized
when using a local cache which means that a very short TTL can be used
without increasing the load on the cluster's DNS resolver. A short TTL
prevents caching in the host's DNS system and hence, removes the need
for explicit host-dependent DNS flushing. Instead, the local cache
is flushed when needed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>