Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Replace LookupHost with a LookupDNS that can handle several query types

The traffic manager's `LookupHost` is only capable of doing DNS queries
of type `A` and `AAAA` (well, in fact it doesn't know, it just resolves
host names to IP addresses). This commit introduces `LookupDNS`, a more
capable lookup that, in addtion to `A`, and `AAAA`, also resolves
queries of type `PTR`, `CNAME`, `MX`, `NS`, SRV`, and `TXT`.

The `LookupHost` is retained for backward compatibility, and will also
be the fallback if a client or agent detects an older traffic-manager.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>