Establish that the user-daemon manages the root-daemon
There's no need for the client ever talk to the root daemon over gRPC.
In particular, it's confusing that the CLI is the one responsible for
doing connect and disconnect to the root daemon when it's in fact the
user-daemon session that is doing all the talking in between.

This commit ensures that the user-daemon takes care of all communication
with the root daemon. The actual start of the process remains in the CLI
though, because the root daemon must be started with admin privileges.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>