Work around for golang lookup glitch
The `net.LookupIP` and `net.LookupCNAME` function makes weird choices
when it comes to search paths. If the name ends with a dot, then no
search path will be applied. If however, the name doesn't end with a
dot, the search path is always applied and the name is never used
verbatim. Both conditions sometimes result in "server misbehaving"
errors.

This commit makes a best effort to guess whether a search path should
be applied, and if the guess isn't certain, falls back to including the
dot in a retry.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>