Get rid of all [wW]ith[Started]Connector and WithRootDaemon
The logic for how the daemons are made accessible to a CLI command has
"evolved" over time until it reached a state that was more or less
incomprehensible, with one `WithConnector`, two `withConnector` in
different packages that performed slightly different tasks,
`WithStartedConnector`, and one `WithRootDaemon`.

This commit therefore replaces all of the above with one `InitCommand`
function that will ensure that the correct things are started and
connected to based on `cobra.Command.Annotations` that it finds in
the command.

The commit also introduces a `need_root_daemon` to the `ConnectRequest`,
because several commands can be executed without having a network
present. In particular, why start the root daemon in order to run
helm commands?

Signed-off-by: Thomas Hallgren <thomas@datawire.io>